refactor: break down WhitelistOriginsCard
